Move base Minecraft events to separate folder
